+++ CHANGELOG.md
@@ -21,6 +21,13 @@ The REST API versions in the table are just for your information as the values a
 
 -   Treemap and Heatmap visualization now emits drill events with `value` property of type `string` instead of `number` same as other visualizations. (BB-1318)
 
+-   Table visualization now emits drill events with correct intersection containing also non-empty `header.identifier` properties. 
+
+
+TODO:
+
+-   Table visualization now emits drill event with `value` property same as other visualizations. (BB-1318)
+
 
 ## 6.2.0